Introduce a base class to manage arbitary supplement files
