+++ npm_and_yarn/lib/dependabot/npm_and_yarn/file_updater.rb
@@ -3,6 +3,7 @@
 require "dependabot/file_updaters"
 require "dependabot/file_updaters/base"
 require "dependabot/file_updaters/vendor_updater"
+require "dependabot/file_updaters/supplement_updater"
 require "dependabot/npm_and_yarn/dependency_files_filterer"
 require "dependabot/npm_and_yarn/sub_dependency_files_filterer"
 
@@ -63,11 +64,11 @@ module Dependabot
 
       def vendor_updated_files(updated_files)
         base_dir = updated_files.first.directory
-        pnp_updater.updated_vendor_cache_files(base_directory: base_dir).each do |file|
+        pnp_updater.updated_files(base_directory: base_dir).each do |file|
           updated_files << file if file.name == ".pnp.cjs" || file.name == ".pnp.data.json"
         end
         vendor_updater.updated_vendor_cache_files(base_directory: base_dir).each { |file| updated_files << file }
-        install_state_updater.updated_vendor_cache_files(base_directory: base_dir).each do |file|
+        install_state_updater.updated_files(base_directory: base_dir).each do |file|
           updated_files << file
         end
 
@@ -95,16 +96,16 @@ module Dependabot
       end
 
       def install_state_updater
-        Dependabot::FileUpdaters::VendorUpdater.new(
+        Dependabot::FileUpdaters::SupplementUpdater.new(
           repo_contents_path: repo_contents_path,
-          vendor_dir: install_state_path
+          target_directory: install_state_path
         )
       end
 
       def pnp_updater
-        Dependabot::FileUpdaters::VendorUpdater.new(
+        Dependabot::FileUpdaters::SupplementUpdater.new(
           repo_contents_path: repo_contents_path,
-          vendor_dir: "./"
+          target_directory: "./"
         )
       end
